requirements.txt
Flask==2.2.3
langchain==0.1.9
langchain-openai==0.0.5
langchain-community==0.0.29
python-dotenv==1.0.0
unstructured==0.16.23
tiktoken==0.4.0
faiss-cpu==1.7.4
libmagic==1.0
python-magic==0.4.27
python-magic-bin==0.4.14
openai==1.12.0
requests==2.31.0
urllib3==2.2.0
certifi==2025.1.31